Initial commit
Code has had some excess cruft and commented out bits deleted. Original
script worked, don't know if I've over trimmed for this intial commit.
In general this code is a bit of a shambles and needs a fair bit of
work.

Note to self, don't have children and side projects.